+++ include/Athena/IStreamReader.hpp
@@ -447,7 +447,7 @@ public:
     {
         vector.clear();
         vector.reserve(count);
-        for (int i=0 ; i<count ; ++i)
+        for (size_t i=0 ; i<count ; ++i)
             vector.emplace_back(readVal<T>());
     }
 
@@ -460,7 +460,7 @@ public:
     {
         vector.clear();
         vector.reserve(count);
-        for (int i=0 ; i<count ; ++i)
+        for (size_t i=0 ; i<count ; ++i)
         {
             vector.emplace_back();
             vector.back().read(*this);
@@ -472,7 +472,7 @@ public:
     {
         vector.clear();
         vector.reserve(count);
-        for (int i=0 ; i<count ; ++i)
+        for (size_t i=0 ; i<count ; ++i)
         {
             vector.emplace_back();
             readf(*this, vector.back());
